Fix include next issue #523
This fixes an issue with #include_next specifically with the ability to
include files that are in the same directory as the compiled one, this
is an error and is solved by simply checking if the path has been
completed on its go through, if it has been, return that we have an
error (empty filename). I was originally going to do this
overcomplicated error handling mechanism when I realized that:

1. It was way too complex for anything useful

2. It was far too much effort

3. It would not change the root of the problem, that which the original
   compiled directory was somehow being included despite #include_next
   by definition skipping it.

These 8 lines fix points 1, 2, and 3.