Fix all problems I know how to fix
The current issue that I cannot fix is the searching from the path of
the last included file, because with files that go into deeper
directories, this functionality is 100% broken, not that it probably
wouldn't cause problems down the line, due to it skipping possibly
necessary imports and seriously messes with anything related to
\#include_next or any functionality that decides to respect the ordering
of include paths, which this absolutely, totally, completely, destroys.

Removing this functionality is also a breaking change, so if we're going
to do this, I would recommend doing this now, when it rears the problems
in implementing #include_next and anything that relies on system files.