Fixes the problem related to include path ordering
I've fixed this bug... for now... presumably...
I don't know how many more times I'll have to change one singular file
in order to perfect #include_next, but I know this will not be the last
time in my life.

However, this should be the last one before I think the major changes I
was talking about need to be considered, I may be able to think of a
solution before doing it for a few more edge cases, but we're hitting a
point that it's becoming near-impossible to determine how many more
edge-cases I can solve.

The #include_next functionality still has more investigating to do to
fully flesh out its edge cases, and we're hitting a point where we
should probably start working on other bugs and things that need to be
added first so that we can find the #include_next bugs.