Add current changes towards a unified template system
Removes the now empty .cpp files
Adds a specific forward declaration section
Make notes for the purpose of most of the compilation
Note that this doesn't compile on MSVC and the MSVC error reporting is gabrage so I need to get this to work on another compiler before I can say what's actually going on.
For the once in the ppExpr I believe it's fine because we only ever create that instance because no other subsystem uses Once objects. This likely should be changed, but I don't know the best way to fix that at this time without further analysis.

Next steps:
Namespacing oasm and ocpp so that the enum class kw doesn't conflict.
Dealing with those ocpp msvc errors I'm trying to dig through.
And in general dealing with a bunch of random non-#embed related action-items that are needed because I just want to use a tokenizer.