Fix some errors and crashes.
Templatereduce.cpp change TemplateDeduceArgList to actually check if the nested type has any data before getting any values out of it.

omake make it show where the command that failed is, should be made to be togglable and should ignore recursive make calls but I added it to track down where errors happen in massive parallel builds.

ppEmbed needed some changes simply to get it to a stage where it only really complains about <variant>

Compiler.h has added FALLTHROUGH because C++17 mandates warning about that, so deal with it.